Pull IntroPanel into its own file
